+++ tests/fixtures.py
@@ -9,8 +9,12 @@ TEST_IMAGE_PATH = os.path.join(TEST_FILES_PATH, "001.dcm")
 TEST_EP2D_IMAGE_PATH = os.path.join(TEST_FILES_PATH, "ep2d_image.dcm")
 TEST_RSFMRI_SERIES_PATH = os.path.join(TEST_FILES_PATH, "rsfmri")
 TEST_RSFMRI_IMAGE_PATH = os.path.join(TEST_FILES_PATH, "rsfmri", "001.dcm")
-TEST_RSFMRI_IMAGE_VOLUME = os.path.join(TEST_FILES_PATH, "rsfmri", "volume.npy")
-TEST_RSFMRI_SERIES_NIFTI = os.path.join(TEST_RSFMRI_SERIES_PATH, "converted.nii.gz")
+TEST_RSFMRI_IMAGE_VOLUME = os.path.join(
+    TEST_FILES_PATH, "rsfmri", "volume.npy"
+)
+TEST_RSFMRI_SERIES_NIFTI = os.path.join(
+    TEST_RSFMRI_SERIES_PATH, "converted.nii.gz"
+)
 TEST_GE_LOCALIZER_PATH = os.path.join(TEST_FILES_PATH, "GE_localizer.dcm")
 TEST_SERIES_PATH = os.path.join(TEST_FILES_PATH, "series")
 TEST_UTILS_DIRECTORY = os.path.join(TESTS_DIRECTORY, "utils")
@@ -22,6 +26,12 @@ TEST_STUDY_FIELDS = {
     "date": datetime.strptime("20180501", "%Y%m%d").date(),
     "time": datetime.strptime("12:21:56.958000", "%H:%M:%S.%f").time(),
 }
+
+SERIES_INSTANCE_UID = (
+    "1.3.12.2.1107.5.2.43.66024.2018050112250992296484473.0.0.0"
+)
+SOP_INSTANCE_UID = "1.3.12.2.1107.5.2.43.66024.2018050112252318571884482"
+STUDY_INSTANCE_UID = "1.3.12.2.1107.5.2.43.66024.30000018050107081466900000007"
 
 
 class ChoiceEnumDefinition(ChoiceEnum):