+++ tests/test_header.py
@@ -1,21 +1,30 @@
+import json
 import warnings
 from pathlib import Path
 from unittest import TestCase
 
 import pydicom
-
 from dicom_parser.header import Header
 from dicom_parser.utils.sequence_detector import SequenceDetector
-from tests.fixtures import (TEST_GE_LOCALIZER_PATH, TEST_IMAGE_PATH,
-                            TEST_STUDY_FIELDS)
+from dicom_parser.utils.value_representation import ValueRepresentation
+
+from tests.fixtures import (
+    SERIES_INSTANCE_UID,
+    SOP_INSTANCE_UID,
+    STUDY_INSTANCE_UID,
+    TEST_GE_LOCALIZER_PATH,
+    TEST_IMAGE_PATH,
+    TEST_RSFMRI_IMAGE_PATH,
+    TEST_STUDY_FIELDS,
+)
 
 
 class HeaderTestCase(TestCase):
     KEYWORDS = {
         "PatientID": "012345678",
-        "SeriesInstanceUID": "1.3.12.2.1107.5.2.43.66024.2018050112250992296484473.0.0.0",
-        "SOPInstanceUID": "1.3.12.2.1107.5.2.43.66024.2018050112252318571884482",
-        "StudyInstanceUID": "1.3.12.2.1107.5.2.43.66024.30000018050107081466900000007",
+        "SeriesInstanceUID": SERIES_INSTANCE_UID,
+        "SOPInstanceUID": SOP_INSTANCE_UID,
+        "StudyInstanceUID": STUDY_INSTANCE_UID,
         "StudyDate": "20180501",
     }
     NON_KEYWORDS = ["ABC", "DEF", "GHI", "JKL"]
@@ -39,12 +48,29 @@ class HeaderTestCase(TestCase):
         ("2222", "3333"),
         ("3333", "4444"),
     ]
-    BAD_DATA_ELEMENT_QUERY_VALUES = 0, ["1"], 1.1, False
+    BAD_DATA_ELEMENT_QUERY_VALUES = (
+        0,
+        ["1"],
+        1.1,
+        False,
+        {"a": "b"},
+        {1, 2, 3},
+        None,
+    )
+    KEYWORD_CONTAINS = {"time": 10, "DATE": 7, "abcdef": 0}
 
     @classmethod
     def setUpClass(cls):
         cls.raw = pydicom.dcmread(TEST_IMAGE_PATH, stop_before_pixels=True)
         cls.header = Header(cls.raw)
+
+    def test_str(self):
+        value = str(self.header)
+        self.assertIsInstance(value, str)
+
+    def test_repr(self):
+        value = repr(Header(TEST_RSFMRI_IMAGE_PATH))
+        self.assertIsInstance(value, str)
 
     def test_instantiation_with_filedataset(self):
         self.assertIsInstance(self.header.raw, pydicom.FileDataset)
@@ -70,41 +96,21 @@ class HeaderTestCase(TestCase):
     def test_init_detected_sequence(self):
         self.assertEqual(self.header.detected_sequence, "Localizer")
 
-    # def test_get_element_by_keyword(self):
-    #     for keyword in self.KEYWORDS.keys():
-    #         result = self.header.get_element_by_keyword(keyword)
-    #         self.assertIsInstance(result, pydicom.DataElement)
+    def test_get_raw_element(self):
+        keys = list(self.TAGS.keys()) + list(self.KEYWORDS.keys())
+        for key in keys:
+            result = self.header.get_raw_element(key)
+            self.assertIsInstance(result, pydicom.DataElement)
 
-    # def test_get_element_by_keyword_with_invalid_key_raises_key_error(self):
-    #     for keyword in self.NON_KEYWORDS:
-    #         with self.assertRaises(KeyError):
-    #             self.header.get_element_by_keyword(keyword)
+    def test_get_raw_value_with_bad_type_raises_type_error(self):
+        for invalid_key in self.BAD_DATA_ELEMENT_QUERY_VALUES:
+            with self.assertRaises(TypeError):
+                self.header.get_data_element(invalid_key)
 
-    # def test_get_element_by_tag(self):
-    #     for tag in self.TAGS.keys():
-    #         result = self.header.get_element_by_tag(tag)
-    #         self.assertIsInstance(result, pydicom.DataElement)
-
-    # def test_get_element_by_tag_invalid_tag_returns_none(self):
-    #     for invalid_tag in self.NON_TAGS:
-    #         with self.assertRaises(KeyError):
-    #             self.header.get_element_by_tag(invalid_tag)
-
-    # def test_get_element(self):
-    #     keys = list(self.TAGS.keys()) + list(self.KEYWORDS.keys())
-    #     for key in keys:
-    #         result = self.header.get_element(key)
-    #         self.assertIsInstance(result, pydicom.DataElement)
-
-    # def test_get_element_with_invalid_key_or_tag_raises_key_error(self):
-    #     for key in self.NON_TAGS + self.NON_KEYWORDS:
-    #         with self.assertRaises(KeyError):
-    #             self.header.get_element(key)
-
-    # def test_get_element_with_non_string_or_tuple_raises_type_error(self):
-    #     for key in self.BAD_DATA_ELEMENT_QUERY_VALUES:
-    #         with self.assertRaises(TypeError):
-    #             self.header.get_element(key)
+    def test_get_data_element_with_bad_type_raises_type_error(self):
+        for invalid_key in self.BAD_DATA_ELEMENT_QUERY_VALUES:
+            with self.assertRaises(TypeError):
+                self.header.get_raw_element(invalid_key)
 
     def test_get_raw_value(self):
         keys = list(self.TAGS.keys()) + list(self.KEYWORDS.keys())
@@ -213,6 +219,15 @@ class HeaderTestCase(TestCase):
         result = header.get("CSASeriesHeaderInfo")
         self.assertIsNone(result)
 
+    def test_get_as_json(self):
+        keys = list(self.TAGS.keys()) + list(self.KEYWORDS.keys())
+        for key in keys:
+            json_value = self.header.get(key, as_json=True)
+            self.assertIsInstance(json_value, str)
+            # Try to execute json.loads() just to make sure no exception is
+            # raised.
+            _ = json.loads(json_value)
+
     def test_detect_sequence(self):
         result = self.header.detect_sequence()
         expected = "Localizer"
@@ -225,3 +240,61 @@ class HeaderTestCase(TestCase):
             warnings.simplefilter("ignore")
             result = header.detect_sequence()
         self.assertIsNone(result)
+
+    def test_keyword_contains(self):
+        for substring, expected in self.KEYWORD_CONTAINS.items():
+            df = self.header.keyword_contains(substring)
+            self.assertEqual(len(df), expected)
+
+    def test_keyword_contains_exact(self):
+        for substring, _ in self.KEYWORD_CONTAINS.items():
+            df = self.header.keyword_contains(substring, exact=True)
+            self.assertEqual(len(df), 0)
+        for substring, expected in self.KEYWORD_CONTAINS.items():
+            df = self.header.keyword_contains(substring.title())
+            self.assertEqual(len(df), expected)
+
+    def test_get_data_elements_with_vr_list(self):
+        vr = [ValueRepresentation.IS, ValueRepresentation.CS]
+        result = self.header.get_data_elements(value_representation=vr)
+        self.assertIsInstance(result, list)
+        self.assertEqual(len(result), 20)
+
+    def test_get_data_elements_with_vr_tuple(self):
+        vr = ValueRepresentation.DA, ValueRepresentation.TM
+        result = self.header.get_data_elements(value_representation=vr)
+        self.assertIsInstance(result, list)
+        self.assertEqual(len(result), 13)
+
+    def test_get_data_elements_with_exclude_list(self):
+        exclude = [
+            ValueRepresentation.IS,
+            ValueRepresentation.CS,
+            ValueRepresentation.DS,
+            ValueRepresentation.UN,
+        ]
+        result = self.header.get_data_elements(exclude=exclude)
+        self.assertIsInstance(result, list)
+        self.assertEqual(len(result), 59)
+
+    def test_get_data_elements_with_exclude_tuple(self):
+        exclude = ValueRepresentation.DA, ValueRepresentation.TM
+        result = self.header.get_data_elements(exclude=exclude)
+        self.assertIsInstance(result, list)
+        self.assertEqual(len(result), 120)
+
+    def test_to_dict(self):
+        value = self.header.to_dict()
+        self.assertIsInstance(value, dict)
+        self.assertEqual(len(value), 119)
+
+    def test_as_dict(self):
+        value = self.header.as_dict
+        expected = self.header.to_dict()
+        self.assertIsInstance(value, dict)
+        self.assertDictEqual(value, expected)
+
+    def test_keys(self):
+        value = self.header.keys
+        self.assertIsInstance(value, list)
+        self.assertEqual(len(value), 119)